+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 13.0.0
 
+### Enhancements
+- **Survey Changes**
+  - Improve survey stability and submission of responses to Slice API
+
 ## 12.0.0 (October 17, 2018)
 
 ### Enhancements